+++ Model/UserDetails.java
@@ -3,7 +3,7 @@ import java.util.Date;
 
 
 public class UserDetails {
-    public int AccountNumber;
+    public long AccountNumber;
     public String FirstName;
     public String LastName;
     public Address addressofUser;
